backend/services/content-processing/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
redis==5.0.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Medical NER and NLP
numpy==1.24.3
spacy==3.6.1
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_md-0.5.3.tar.gz

# Text Processing
nltk==3.8.1
textstat==0.7.3

# AWS and Storage
boto3==1.34.0

# Event handling
kafka-python==2.0.2

# Security & Auth
python-jose[cryptography]==3.3.0
httpx==0.25.2

# Utilities
structlog==23.2.0
python-magic==0.4.27